+++ Projects/Features/MainFeature/BroadcastUploadExtension/SampleHandler.swift
@@ -1,0 +1,6 @@
+import ReplayKit
+
+final class SampleHandler: RPBroadcastSampleHandler, @unchecked Sendable {
+    override func broadcastStarted(withSetupInfo setupInfo: [String: NSObject]?) {
+    }
+}
